+++ packages/transform-eagle/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.12](https://github.com/rao-pics/core/compare/@raopics/transform-eagle@1.3.11...@raopics/transform-eagle@1.3.12) (2023-04-08)
+
+**Note:** Version bump only for package @raopics/transform-eagle
+
 ## [1.3.11](https://github.com/rao-pics/core/compare/@raopics/transform-eagle@1.3.10...@raopics/transform-eagle@1.3.11) (2023-04-04)
 
 **Note:** Version bump only for package @raopics/transform-eagle