+++ packages/use/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.10](https://github.com/rao-pics/core/compare/@raopics/use@0.3.9...@raopics/use@0.3.10) (2023-04-08)
+
+**Note:** Version bump only for package @raopics/use
+
 ## [0.3.9](https://github.com/rao-pics/core/compare/@raopics/use@0.3.8...@raopics/use@0.3.9) (2023-04-04)
 
 **Note:** Version bump only for package @raopics/use